Cithara/requirements.txt
Django>=4.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
requests>=2.28.0
python-dotenv>=1.0.0
